+++ Source/Scene/OrthographicFrustum.js
@@ -5,6 +5,7 @@ define([
         '../Core/Cartesian4',
         '../Core/defined',
         '../Core/defineProperties',
+        '../Core/deprecationWarning',
         '../Core/DeveloperError',
         '../Core/Matrix4',
         './CullingVolume'
@@ -14,6 +15,7 @@ define([
         Cartesian4,
         defined,
         defineProperties,
+        deprecationWarning,
         DeveloperError,
         Matrix4,
         CullingVolume) {
@@ -270,7 +272,7 @@ define([
      * Returns the pixel's width and height in meters.
      *
      * @param {Cartesian2} drawingBufferDimensions A {@link Cartesian2} with width and height in the x and y properties, respectively.
-     * @param {Number} [distance=near plane distance] The distance to the near plane in meters.
+     * @param {Number} distance Ignored for orthographic frustum.
      * @param {Cartesian2} [result] The object onto which to store the result.
      * @returns {Cartesian2} The modified result parameter or a new instance of {@link Cartesian2} with the pixel's width and height in the x and y properties, respectively.
      *
@@ -278,11 +280,13 @@ define([
      * @exception {DeveloperError} drawingBufferDimensions.y must be greater than zero.
      *
      * @example
-     * // Example 1
-     * // Get the width and height of a pixel.
      * var pixelSize = camera.frustum.getPixelSize(new Cesium.Cartesian2(canvas.clientWidth, canvas.clientHeight));
+     *
+     * @deprecated
      */
     OrthographicFrustum.prototype.getPixelSize = function(drawingBufferDimensions, distance, result) {
+        deprecationWarning('OrthographicFrustum', 'getPixelSize is deprecated. Use getPixelDimensions instead.');
+
         update(this);
 
         //>>includeStart('debug', pragmas.debug);
@@ -305,6 +309,54 @@ define([
         if (!defined(result)) {
             return new Cartesian2(pixelWidth, pixelHeight);
         }
+
+        result.x = pixelWidth;
+        result.y = pixelHeight;
+        return result;
+    };
+
+    /**
+     * Returns the pixel's width and height in meters.
+     *
+     * @param {Number} drawingBufferWidth The width of the drawing buffer.
+     * @param {Number} drawingBufferHeight The height of the drawing buffer.
+     * @param {Number} distance The distance to the near plane in meters.
+     * @param {Cartesian2} result The object onto which to store the result.
+     * @returns {Cartesian2} The modified result parameter or a new instance of {@link Cartesian2} with the pixel's width and height in the x and y properties, respectively.
+     *
+     * @exception {DeveloperError} drawingBufferWidth must be greater than zero.
+     * @exception {DeveloperError} drawingBufferHeight must be greater than zero.
+     *
+     * @example
+     * // Example 1
+     * // Get the width and height of a pixel.
+     * var pixelSize = camera.frustum.getPixelDimensions(canvas.clientWidth, canvas.clientHeight, 0.0, new Cartesian2());
+     */
+    OrthographicFrustum.prototype.getPixelDimensions = function(drawingBufferWidth, drawingBufferHeight, distance, result) {
+        update(this);
+
+        //>>includeStart('debug', pragmas.debug);
+        if (!defined(drawingBufferWidth) || !defined(drawingBufferHeight)) {
+            throw new DeveloperError('Both drawingBufferWidth and drawingBufferHeight are required.');
+        }
+        if (drawingBufferWidth <= 0) {
+            throw new DeveloperError('drawingBufferWidth must be greater than zero.');
+        }
+        if (drawingBufferHeight <= 0) {
+            throw new DeveloperError('drawingBufferHeight must be greater than zero.');
+        }
+        if (!defined(distance)) {
+            throw new DeveloperError('distance is required.');
+        }
+        if (!defined(result)) {
+            throw new DeveloperError('A result object is required.');
+        }
+        //>>includeEnd('debug');
+
+        var frustumWidth = this.right - this.left;
+        var frustumHeight = this.top - this.bottom;
+        var pixelWidth = frustumWidth / drawingBufferWidth;
+        var pixelHeight = frustumHeight / drawingBufferHeight;
 
         result.x = pixelWidth;
         result.y = pixelHeight;